projects/humanlike-agent/requirements.txt
telethon==1.*
aiohttp==3.*
python-dotenv==1.*
numpy==2.*
